refactor(config): move parseSkillSelection to the module that owns the column
Structure only, no behavior change.

Why: the parser reads `container_configs.skills`. That column belongs to
container-config.ts, which is also where the second, untolerant reading of it
lives — keeping the parser one import away in the composer is what let the two
drift apart.

What: `parseSkillSelection` moves from `src/project-doc-compose.ts` to
`src/container-config.ts`; the composer and `container-runner.ts` import it
from there. No cycle: the composer already imports `sanitizeStoredMcpServers`
from the same module.

Verification: `pnpm exec tsc --noEmit` clean; `pnpm exec vitest run
src/project-doc-compose.test.ts src/container-config.test.ts
src/container-runner*.test.ts` 106 pass, unchanged.
